requirements.txt
Django==3.1.6
psycopg2
tensorflow>=1.14.0
keras
pandas
numpy
gensim
mecab-python3==0.996.2